+++ src/main/resources/db/migration/V7__create_drink_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE drink (
     id INT PRIMARY KEY,
     volume INT,
     alcoolico BOOLEAN,
-    teor_alcoolico NUMERIC(5, 2),
+    teor_alcoolico FLOAT,
     gaseificada BOOLEAN,
     FOREIGN KEY (id) REFERENCES item (id) ON DELETE CASCADE
 );
